# tests/Swagger_updated_20-10-2017.yaml
swagger: '2.0'
info:
  version: '2.0'
  title: "API Project I'm Ready"
  description: "API Project I'm Ready"
# basePath: "/"
# host: "imready.ml"
...
produces:
  - "application/json"
schemes:
 - "https"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...
definitions:
...
  Caregivers:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
        description: "Name of the user"
      role:
        description: "Role of client"
        type: "string"
        enum:
        - "admin"
        - "client"
        - "caregiver"
        - "family"
      messages:
        description: "All messages this user has sent or received"
        type: "array"
        items: {
          $ref: "#/definitions/Messages"
        }        
      appointments:
        description: "All appointments of the user"
        type: "array"
        items: {
          $ref: "#/definitions/Appointments"
        }
  EntityModel:
    type: "object"
    properties:
      id:
        type: "string"
...
  Appointments:
    type: "object"
    properties:
      id:
        description: "ID of the appointment"
        type: "string"
      datetime:
        description: "Deadline of the appointment"
        type: "string"
        format: "datetime"
      description:
        description: "Description of the appointment"
        type: "string"    
  InputMessages:
    type: "object"
    properties:
      receiverId:
        description: "Receiver ID of the message"
        type: "string"
      message:
        description: "Content of the message"
        type: "string"
  Messages:
    type: "object"
    properties:
      id:
        description: "ID of the message"
        type: "string"
      senderId:
        description: "ID of the sender of the message"
        type: "string"
      receiverId:
        description: "ID of the receiver of the message"
        type: "string"
      message:
        description: "Content of the message"
        type: "string"
      datetime:
        description: "Time the message was sent"
        type: "string"
        format: "datetime"
      read:
        description: "The read status of the message"
        type: "boolean"
...
  Credentials:
    type: "object"
    properties:
      username: 
        description: "Username"
        type: "string"
      password:
        description: "Password"
        type: "string"
...
  ErrorResponses:
    type: "object"
    properties:
      code:
        description: "Error code"
        type: "integer"
        format: "int64"
      message:
        description: "Error message"
        type: "string"        
paths:
...
  /caregivers:
    get:
      summary: "Get caregivers"
      description: "Returns a list of caregivers, starting with the newest caregivers in the database that fits the specified arguments."
      tags:
      - "general"
      parameters:
      - name: "count"
        in: "query"
        description: "The maximum number of caregivers to return."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "succesful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    post:
      summary: "Create a new caregiver"
      description: "Create a new caregiver"
      tags:
      - "general"
      consumes:
      - "application/json"
      parameters:
      - name: "Credentials"
        in: "body"
        description: "Credentials of the user"
        required: true
        schema:
          $ref: "#/definitions/Credentials" 
      responses:
        201:
          description: "succesful created new caregiver"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        409:
          description: "Username already exists"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
  /caregivers/{id}:
    get:
      summary: "Get caregiver"
      description: "Returns a caregiver that fits the specified arguments."
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of caregiver to return"
        required: true
        type: "string"
      responses:
        200:
          description: "succesful operation"
          schema:
            $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update caregiver"
      description: "Update caregiver with specified ID"
      tags:
      - "general"
      consumes: 
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of caregiver to be updated"
        required: true
        type: "string"
      - name: "Credentials"
        in: "body"
        description: "Credentials of the user"
        required: true
        schema:
          $ref: "#/definitions/Credentials"
      responses:
        200:
          description: "succesfully updated caregiver"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete caregiver"
      description: "Delete caregiver with specified ID"
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the caregiver to deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "succesfully deleted caregiver"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
  /clients/{clientId}/caregivers/{caregiverId}:
    get:
      summary: "Get a single client's caregiver with ID"
      description: "Returns a single caregiver with specified userID and caregiverID"
      tags:
        - "connections"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be returned"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be returned"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfully found the caregigver by ID"
          schema:
            $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client or caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update client's caregiver"
      description: "Update caregiver with specified userID and caregiverID"
      tags:
      - "connections"
      consumes:
      - "application/json"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be updated"
        required: true
        type: "string"
      - name: "caregiver"
        in: "body"
        description: "Updated caregiver object"
        required: true
        schema:
          $ref: "#/definitions/EntityModel"
      responses:
        200:
          description: "Succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client or caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete client's caregiver"
      description: "Delete client's caregiver with specified ID"
      tags:
      - "connections"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "Succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
  /family/{familyId}/messages/{messageId}:
    get:
      summary: "Get specific family message"
      description: "Get family message"
      tags:
      - "messages"
      parameters:
      - name: "familyId"
        in: "path"
        description: "ID of the family"
        required: true
        type: "string"
      - name: "messageId"
        in: "path"
        description: "ID of the message to be returned"
        required: true
        type: "string"
      responses:
        200:
          description: "Successfully returned the messages"
          schema:
            $ref: "#/definitions/Messages"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete specific family message"
      description: "delete family message"
      tags:
      - "messages"
      parameters:
      - name: "familyId"
        in: "path"
        description: "ID of the family"
        required: true
        type: "string"
      - name: "messageId"
        in: "path"
        description: "ID of the message to be returned"
        required: true
        type: "string"
      responses:
        204:
          description: "Successfully deleted the messages"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
  /clients/{clientId}/messages/{messageId}:
    get:
      summary: "Get specific client message"
      description: "Get client message"
      tags:
      - "messages"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of the client"
        required: true
        type: "string"
      - name: "messageId"
        in: "path"
        description: "ID of the message to be returned"
        required: true
        type: "string"
      responses:
        200:
          description: "Successfully returned the messages"
          schema:
            $ref: "#/definitions/Messages"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete specific client message"
      description: "delete client message"
      tags:
      - "messages"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of the family"
        required: true
        type: "string"
      - name: "messageId"
        in: "path"
        description: "ID of the message to be returned"
        required: true
        type: "string"
      responses:
        204:
          description: "Successfully deleted the messages"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
  /caregivers/{caregiverId}/messages/:
    get:
      summary: "Get all caregiver messages"
      description: "Get all caregiver messages"
      tags:
      - "messages"
      parameters:
      - name: "caregiverId"
        in: "path"
        required: true
        type: "string"
        description: "Needs an Id to identify the caregiver."
      - name: "count"
        in: "query"
        description: "The maximum number of messages to return."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successfully returned the messages"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Messages"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    post:
      summary: "Send a new message to a caregivers"
      description: "Add a new message"
      tags:
      - "messages"
      consumes:
      - "application/json"
      parameters:
      - name: "caregiverId"
        in: "path"
        required: true
        type: "string"
        description: "Needs an Id to identify the caregiver."
      - name: "message"
        in: "body"
        description: "Message object"
        required: true
        schema:
          $ref: "#/definitions/InputMessages"
      responses:
        201:
          description: "The message has been succesfully sent"
        400:
          description: "The input parameter(s) are invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "An unauthorized attempt"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
  /caregivers/{caregiverId}/messages/{messageId}:
    get:
      summary: "Get specific family message"
      description: "Get family message"
      tags:
      - "messages"
      parameters:
      - name: "caregiverId"
        in: "path"
        description: "ID of the caregiver"
        required: true
        type: "string"
      - name: "messageId"
        in: "path"
        description: "ID of the message to be returned"
        required: true
        type: "string"
      responses:
        200:
          description: "Successfully returned the messages"
          schema:
            $ref: "#/definitions/Messages"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete specific caregiver message"
      description: "delete caregiver message"
      tags:
      - "messages"
      parameters:
      - name: "caregiverId"
        in: "path"
        description: "ID of the caregiver"
        required: true
        type: "string"
      - name: "messageId"
        in: "path"
        description: "ID of the message to be returned"
        required: true
        type: "string"
      responses:
        204:
          description: "Successfully deleted the messages"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /InhollandCloudAPI/ProjectAPI/2.0